% Auto-generated: do not edit by hand
\name{nX}

\alias{nX}

\title{NX component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
nX(id=NULL, data=NULL, tapNode=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Named list. Vega-lite specification of the plot.}

\item{tapNode}{Named list. Node clicked.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

